requirements.txt
setuptools>=19.0
dendropy>=4.1
pystache>=0.5.4
rdflib>=4.2.1
pytest-pylint
